Move static methods from wrangler classes into separate `util` modules to prevent namespace cluttering.
