+++ .travis.yml
@@ -10,7 +10,7 @@ python:
 env:
   - ENV_STRING=master
 #  - ENV_STRING=pandas0.24.1
-  - ENV_STRING=pandas0.24.0
+#  - ENV_STRING=pandas0.24.0
 #
 #  - ENV_STRING=pandas0.23.4
 #  - ENV_STRING=pandas0.23.3
@@ -32,10 +32,10 @@ env:
 #  - ENV_STRING=pandas0.19.1
 #  - ENV_STRING=pandas0.19.0
 #
-  - ENV_STRING=pyspark2.4.0
+#  - ENV_STRING=pyspark2.4.0
 #  - ENV_STRING=pyspark2.3.1
 
-  - ENV_STRING=dask1.1.5
+#  - ENV_STRING=dask1.1.5
 
 
 # Remove python/pandas version interactions which do not have wheels on pypi
